requirements.txt
fastapi
uvicorn
pydantic-ai-slim[bedrock]>=0.4.0
boto3>=1.34.0
langgraph>=0.6.0
langchain-core>=0.3.72
tenacity>=9.1.2
